Moved keypad out of LeftPanel
