+++ src/components/LeftPanel.jsx
@@ -7,6 +7,7 @@ import FlatButton from 'material-ui/FlatButton';
 
 import { toggleUI, addOperationChar, goCompute, clearDisplay, monkeyClickAsync,
   monkeyModeActivation, monkeyModeDeactivation } from '../actions';
+import { NumericKeypad } from './NumericKeypad.jsx';
 
 const primaryButtonStyle = {
   width: '80px',
@@ -28,19 +29,8 @@ const fadedButtonStyle = {
 export class LeftPanel extends React.Component {
   constructor() {
     super();
-    this.renderButton = this.renderButton.bind(this);
-    this.getPressedStatus = this.getPressedStatus.bind(this);
     this.handleMonkeyClick = this.handleMonkeyClick.bind(this);
-    this.handleOperationClick = this.handleOperationClick.bind(this);
     this.handleGoClick = this.handleGoClick.bind(this);
-    this.handleClearClick = this.handleClearClick.bind(this);
-  }
-
-  getPressedStatus(keyName) {
-    if (keyName === this.props.keyPressed) {
-      return true;
-    }
-    return false;
   }
 
   handleMonkeyClick() {
@@ -51,84 +41,19 @@ export class LeftPanel extends React.Component {
     }
   }
 
-  handleOperationClick(value) {
-    this.props.onOperationClick(value);
-  }
-
   handleGoClick() {
     this.props.onGoClick();
-  }
-
-  handleClearClick() {
-    this.props.onClearClick();
-  }
-
-  renderButton(keyName) {
-    const isPressed = this.getPressedStatus(keyName);
-    const style = {
-      borderRadius: '25px',
-      backgroundColor: '#ddd',
-      minWidth: '40px',
-      margin: '5px',
-    };
-    if (isPressed) {
-      style.backgroundColor = '#FFB74D';
-    }
-    return (
-      <FlatButton
-        style={style}
-        onClick={() => { this.handleOperationClick(keyName); }}
-      >
-        {keyName}
-      </FlatButton>
-    );
   }
 
   render() {
     return (
       <div id="left-panel">
-        <div id="calc-secondary-keys">
-          <div className="calc-num-container">
-            <div className="calc-num-row">
-              {this.renderButton('1')}
-              {this.renderButton('2')}
-              {this.renderButton('3')}
-            </div>
+        <NumericKeypad
+          handleOperationClick={this.props.onOperationClick}
+          keyPressed={this.props.keyPressed}
+          handleClearClick={this.props.onClearClick}
+        />
 
-            <div className="calc-num-row">
-              {this.renderButton('4')}
-              {this.renderButton('5')}
-              {this.renderButton('6')}
-            </div>
-
-            <div className="calc-num-row">
-              {this.renderButton('7')}
-              {this.renderButton('8')}
-              {this.renderButton('9')}
-            </div>
-            <div className="calc-num-row">
-              {this.renderButton('0')}
-              {this.renderButton('.')}
-              <FlatButton
-                style={{
-                  borderRadius: '25px',
-                  backgroundColor: '#ddd',
-                  minWidth: '40px',
-                  margin: '5px',
-                }}
-                onClick={this.handleClearClick}
-              >
-                c
-              </FlatButton>
-            </div>
-          </div>
-          <div className="calc-operations-bunch">
-            {this.renderButton('+')}
-            {this.renderButton('-')}
-            {this.renderButton('*')}
-            {this.renderButton('/')}
-          </div>
-        </div>
         <div style={{ display: 'flex', flexDirection: 'row', marginBottom: '5px' }}>
           <FlatButton
             onClick={this.handleGoClick}
